src/backend/requirements.txt
langchain==0.0.274
gpt4all==1.0.8
pinecone-client==2.2.2
llama-cpp-python==0.1.81
urllib3==2.0.4
PyMuPDF==1.23.5
python-dotenv==1.0.0
unstructured==0.10.8
extract-msg==0.45.0
tabulate==0.9.0
pandoc==2.3
pypandoc==1.11
tqdm==4.66.1
sentence-transformers==2.2.2
huggingface-hub<0.17.0
python-multipart==0.0.6


fastapi==0.110.0
uvicorn==0.27.0
pinecone-client==3.0.0
python-dotenv==1.0.0
numpy==1.26.0
pydantic==2.5.0